Add more schema and attribute work
