+++ plugins/filter/as_datetime.yml
@@ -1,0 +1,50 @@
+# Copyright (c) 2023 Ansible Project
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+DOCUMENTATION:
+  name: as_datetime
+  author:
+  - Jordan Borean (@jborean93)
+  short_description: Converts an LDAP value to a datetime string
+  seealso:
+  - filter: microsoft.ad.as_guid
+  - filter: microsoft.ad.as_sid
+  - inventory: microsoft.ad.inventory
+  description:
+  - Converts an LDAP integer or raw value to a datetime string.
+  - Should be used with the C(microsoft.ad.inventory) plugin to convert
+    attribute values to a datetime string.
+  positional: _input
+  options:
+    _input:
+      description:
+      - The LDAP attribute bytes or integer value representing a FILETIME
+        integer stored in LDAP.
+      - The resulting datetime will be set as a UTC datetime as that's how the
+        FILETIME value is stored in LDAP.
+      type: raw
+      required: true
+    format:
+      description:
+      - The string format to format the datetime object as.
+      - Defaults to an ISO 8601 compatible string, for example
+        C(2023-02-06T07:39:09.195321+0000).
+      default: '%Y-%m-%dT%H:%M:%S.%f%z'
+      type: str
+
+EXAMPLES: |
+  # This is an example used in the microsoft.ad.inventory plugin
+
+  # Converting from the coerced value
+  attributes:
+    pwdLastSet: this | microsoft.ad.as_datetime
+
+  # Converting from the raw bytes value
+  attributes:
+    maxPwdAge: raw | microsoft.ad.as_datetime
+
+RETURN:
+  _value:
+    description:
+    - The datetime string value(s) formated as per the I(format) option.
+    type: string
