+++ plugins/filter/as_sid.yml
@@ -1,0 +1,43 @@
+# Copyright (c) 2023 Ansible Project
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+DOCUMENTATION:
+  name: as_sid
+  author:
+  - Jordan Borean (@jborean93)
+  short_description: Converts an LDAP value to a Security Identifier string
+  seealso:
+  - filter: microsoft.ad.as_datetime
+  - filter: microsoft.ad.as_guid
+  - inventory: microsoft.ad.inventory
+  description:
+  - Converts an LDAP string or raw value to a security identifier string.
+  - Should be used with the C(microsoft.ad.inventory) plugin to convert
+    attribute values to a security identifier string.
+  positional: _input
+  options:
+    _input:
+      description:
+      - The LDAP attribute bytes or string value representing a Security
+        Identifier stored in LDAP.
+      - If using a string as input, it must be a base64 string representing
+        the SIDs bytes.
+      type: raw
+      required: true
+
+EXAMPLES: |
+  # This is an example used in the microsoft.ad.inventory plugin
+
+  # Converting from the coerced value
+  attributes:
+    objectSid: this | microsoft.ad.as_sid
+
+  # Converting from the raw bytes value
+  attributes:
+    objectSid: raw | microsoft.ad.as_sid
+
+RETURN:
+  _value:
+    description:
+    - The security identifier string value(s).
+    type: string
